+++ .gitignore
@@ -1,6 +1,7 @@
 HELP.md
 .gradle
 build/
+/src/main/generated/
 !gradle/wrapper/gradle-wrapper.jar
 !**/src/main/**/build/
 !**/src/test/**/build/
