+++ src/main/java/Book/yes48/Entity/goods/form/AdminGoodsSearchDto.java
@@ -1,0 +1,14 @@
+package Book.yes48.Entity.goods.form;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@RequiredArgsConstructor
+public class AdminGoodsSearchDto {
+
+    private String searchBy;
+    private String searchQuery = "";
+}
