+++ src/main/java/Book/yes48/Entity/goods/form/AdminSearchCondition.java
@@ -1,0 +1,39 @@
+package Book.yes48.Entity.goods.form;
+
+import Book.yes48.Entity.FileStore;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter @Setter
+public class AdminSearchCondition {
+
+    private Long id;
+    private String name;
+    private String sort;
+    private String author;
+    private String publisher;
+    private String publisherDate;
+    private Integer price;
+    private Integer stockQuantity;
+    private String event;
+    private String state;
+    private FileStore fileStore;
+    private String searchBy;
+    private String searchQuery = "";
+
+    public AdminSearchCondition(Long id, String name, String sort, String author, String publisher, String publisherDate, Integer price, Integer stockQuantity, String event, String state, FileStore fileStore, String searchBy, String searchQuery) {
+        this.id = id;
+        this.name = name;
+        this.sort = sort;
+        this.author = author;
+        this.publisher = publisher;
+        this.publisherDate = publisherDate;
+        this.price = price;
+        this.stockQuantity = stockQuantity;
+        this.event = event;
+        this.state = state;
+        this.fileStore = fileStore;
+        this.searchBy = searchBy;
+        this.searchQuery = searchQuery;
+    }
+}
