--- src/main/java/com/example/yes48/Entity/order/Order.java
@@ -1,34 +1,0 @@
-package com.example.yes48.Entity.order;
-
-import com.example.yes48.Entity.BaseTimeEntity;
-import com.example.yes48.Entity.member.Member;
-import jakarta.persistence.*;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-
-@Entity
-@Table(name = "orders")
-@Getter
-public class Order extends BaseTimeEntity {
-
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "order_id")
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "member_id")
-    private Member member;
-
-    @OneToMany(mappedBy = "order", cascade = CascadeType.ALL)
-    private List<OrderItem> orderItems = new ArrayList<>();
-
-    private int quantity;
-    private int totalPrice;
-    private LocalDateTime orderDate;
-
-    @Enumerated(EnumType.STRING)
-    private OrderStatus status;
-}
