--- src/main/java/com/example/yes48/Entity/order/OrderStatus.java
@@ -1,6 +1,0 @@
-package com.example.yes48.Entity.order;
-
-public enum OrderStatus {
-
-    ORDER, CANCEL
-}
