--- src/main/java/com/example/yes48/Yes48Application.java
@@ -1,13 +1,0 @@
-package com.example.yes48;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Yes48Application {
-
-	public static void main(String[] args) {
-		SpringApplication.run(Yes48Application.class, args);
-	}
-
-}
