--- src/main/java/com/example/yes48/web/controller/GoodsController.java
@@ -1,19 +1,0 @@
-package com.example.yes48.web.controller;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-
-@Controller
-public class GoodsController {
-
-    @GetMapping("/book-list")
-    public String list() {
-        return "book/list";
-    }
-
-    @GetMapping("/goods")
-    public String detail() {
-        return "book/goods";
-    }
-}
