+++ src/main/resources/application.yml
@@ -9,18 +9,19 @@ spring:
       max-file-size: 10MB
       max-request-size: 10MB
 
-
   jpa:
     generate-ddl: true
     open-in-view: true
     hibernate:
-      ddl-auto: create
+      ddl-auto: none
       naming:
         physical-strategy: org.hibernate.boot.model.naming.PhysicalNamingStrategyStandardImpl
       use-new-id-generator-mappings: false
     show-sql: true
     properties:
-      hibernate.format_sql: true
+      hibernate:
+        show_sql: true
+        format_sql: true
       dialect: org.hibernate.dialect.MySQL8InnoDBDialect
