+++ src/main/resources/templates/layout/adminNavbar.html
@@ -11,13 +11,6 @@
         <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
           <span class="navbar-toggler-icon"></span>
         </button>
-        <div class="collapse navbar-collapse" id="navbarSupportedContent">
-          <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-            <li class="nav-item">
-              <a href="/admin/goods" class="nav-link active" aria-current="page" style="color: #ffffff">상품관리</a>
-            </li>
-          </ul>
-        </div>
       </div>
     </nav>
   </div>
