+++ src/test/java/Book/yes48/Yes48ApplicationTests.java
@@ -1,0 +1,13 @@
+package Book.yes48;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class Yes48ApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
