--- src/test/java/com/example/yes48/Yes48ApplicationTests.java
@@ -1,13 +1,0 @@
-package com.example.yes48;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class Yes48ApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
